+++ src/types/PortmanConfig.ts
@@ -86,6 +86,7 @@ export type ResponseBodyTest = {
   key: string
   value?: string | number | boolean
   contains?: string
+  oneOf?: string[] | number[] | boolean[]
   length?: string | number
   minLength?: string | number
   maxLength?: string | number
@@ -95,6 +96,7 @@ export type ResponseHeaderTest = {
   key: string
   value?: string | number | boolean
   contains?: string
+  oneOf?: string[]
   length?: string | number
   minLength?: string | number
   maxLength?: string | number